Release #2
The background refresh is now running on async instead of a new thread. Now it takes less ressources!. I also added a token to it allowing now to stop the loop for real.
Fixed a bug where the service don't really start at Windows startup (I never set it up lol);
The main form is not frozing anymore when starting and stopping services;
Fixed a bug that occured at Windows startup when we click on the icon in the system tray.