Release 6
The code is a bit more clear with better methods names and by removing useless spaces.
Fixed a bug that used to make multiple searching loop while unchecking the "Run in background" box when Oracle was still on.
Fixed a bug that used to appear when we start SOS while Oracle is already running. The icon in the system tray was red and was supposed to be green.
When  the app is not in Alt + Tab anymore when we minimize it.