Release 8
I don't know how to rename the "ServiceOracle" folder so that's why the title is in French. However the csproj file name is in English and I never edited the sln file. That's why it did not used to work for those who didn't know how to fix it. My bad.

The Try Catch in BackgroundRefresh is now gone (didn't know the usage of the question mark) and I changed some methods names to make it easier to read.