Release 9
The same program but in much better.
The code is so next level compared to before. I almost rewrote it from scratch. In other word, it is so much more easy to read and to edit...
Yeah I had a bad time trying to understand it 😅
It also looks a bit faster and is not freezing anymore. Now with .NET 7.0.
I don't know if the old crappy code is the reason but after compilying it, its size almost got sliced in half.